[tools:publish,replaceinfiles,setup] Build | Automated versioning

+++ tools/publish.sh
@@ -2,6 +2,13 @@
 
 PYTHON=python3.9
 
+if [ -z $VERSION ] ; then
+	echo VERSION env. must be specified!
+	exit 1
+fi
+
+replaceinfiles.py 'setup.py' '@VERSION@' $VERSION
+
 echo starting.. \
 	&& rm ../setup.py \
 	&& cp setup.py ..  \
